* logonserver now supports auth reconnect. you can enter world, jump back to the character list and choose another realm now. :) * decreased memory usage of logonserver (helps on large servers with shitloads of accounts)
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@1745 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/WorldSocket.cpp
@@ -237,7 +237,7 @@ void WorldSocket::InformationRetreiveCallback(WorldPacket & recvData, uint32 req
 	string AccountName;
 	uint32 AccountID;
 	string GMFlags;
-	uint32 AccountFlags;
+	uint8 AccountFlags;
 	
 	recvData >> AccountID >> AccountName >> GMFlags >> AccountFlags;
 	sLog.outDebug( " >> got information packet from logon: `%s` ID %u (request %u)", AccountName.c_str(), AccountID, mRequestID);
